move constants defination to constants.h
